Extract `DrugsSummaryViewHolder` to a separate class
