+++ app/src/main/java/org/simple/clinic/summary/SummaryPrescribedDrugsItem.kt
@@ -1,29 +1,21 @@
 package org.simple.clinic.summary
 
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import com.xwray.groupie.ViewHolder
 import io.reactivex.subjects.Subject
-import kotlinx.android.extensions.LayoutContainer
-import kotlinx.android.synthetic.main.list_patientsummary_prescriptions.*
 import org.simple.clinic.R
 import org.simple.clinic.drugs.PrescribedDrug
 import org.simple.clinic.summary.SummaryListAdapterIds.PRESCRIBED_DRUGS
-import org.simple.clinic.text.style.TextAppearanceWithLetterSpacingSpan
+import org.simple.clinic.summary.prescribeddrugs.DrugsSummaryViewHolder
 import org.simple.clinic.util.RelativeTimestampGenerator
-import org.simple.clinic.util.Truss
 import org.simple.clinic.util.UserClock
 import org.simple.clinic.widgets.UiEvent
-import org.simple.clinic.widgets.visibleOrGone
 import org.threeten.bp.format.DateTimeFormatter
 
 data class SummaryPrescribedDrugsItem(
     val prescriptions: List<PrescribedDrug>,
     val dateFormatter: DateTimeFormatter,
     val userClock: UserClock
-) : GroupieItemWithUiEvents<SummaryPrescribedDrugsItem.DrugsSummaryViewHolder>(adapterId = PRESCRIBED_DRUGS) {
+) : GroupieItemWithUiEvents<DrugsSummaryViewHolder>(adapterId = PRESCRIBED_DRUGS) {
 
   private val relativeTimestampGenerator = RelativeTimestampGenerator()
 
@@ -37,88 +29,5 @@ data class SummaryPrescribedDrugsItem(
 
   override fun bind(holder: DrugsSummaryViewHolder, position: Int) {
     holder.bind(prescriptions, dateFormatter, relativeTimestampGenerator, userClock)
-  }
-
-  class DrugsSummaryViewHolder(
-      override val containerView: View,
-      private val onUpdateClicked: () -> Unit
-  ) : ViewHolder(containerView), LayoutContainer {
-
-    init {
-      updateButton.setOnClickListener { onUpdateClicked() }
-    }
-
-    fun bind(
-        prescriptions: List<PrescribedDrug>,
-        dateFormatter: DateTimeFormatter,
-        timestampGenerator: RelativeTimestampGenerator,
-        userClock: UserClock
-    ) {
-      summaryViewGroup.visibleOrGone(prescriptions.isNotEmpty())
-
-      setButtonText(prescriptions)
-
-      removeAllDrugViews()
-
-      if (prescriptions.isNotEmpty()) {
-        prescriptions.forEach { drug ->
-          val drugViewHolder = inflateRowForDrug()
-          drugViewHolder.bind(drug)
-        }
-
-        val lastUpdatedPrescription = prescriptions.maxBy { it.updatedAt }!!
-
-        val lastUpdatedTimestamp = timestampGenerator.generate(lastUpdatedPrescription.updatedAt, userClock)
-
-        lastUpdatedTimestampTextView.text = itemView.resources.getString(
-            R.string.patientsummary_prescriptions_last_updated,
-            lastUpdatedTimestamp.displayText(itemView.context, dateFormatter)
-        )
-      }
-    }
-
-    private fun inflateRowForDrug(): DrugViewHolder {
-      val drugViewHolder = DrugViewHolder.create(drugsSummaryContainer)
-      drugsSummaryContainer.addView(drugViewHolder.itemView, drugsSummaryContainer.childCount - 1)
-      return drugViewHolder
-    }
-
-    private fun removeAllDrugViews() {
-      drugsSummaryContainer.removeAllViews()
-    }
-
-    private fun setButtonText(prescriptions: List<PrescribedDrug>) {
-      updateButton.text =
-          if (prescriptions.isEmpty()) {
-            itemView.context.getString(R.string.patientsummary_prescriptions_add)
-          } else {
-            itemView.context.getString(R.string.patientsummary_prescriptions_update)
-          }
-    }
-
-    class DrugViewHolder(val itemView: View) {
-      private val drugTextView = itemView as TextView
-
-      fun bind(drug: PrescribedDrug) {
-        val summaryBuilder = Truss()
-        summaryBuilder.append(drug.name)
-        if (drug.dosage.isNullOrBlank().not()) {
-          val dosageTextAppearance = TextAppearanceWithLetterSpacingSpan(itemView.context, R.style.Clinic_V2_TextAppearance_Body1Left_Grey1)
-          summaryBuilder
-              .pushSpan(dosageTextAppearance)
-              .append("  ${drug.dosage}")
-              .popSpan()
-        }
-        drugTextView.text = summaryBuilder.build()
-      }
-
-      companion object {
-        fun create(parent: ViewGroup): DrugViewHolder {
-          val inflater = LayoutInflater.from(parent.context)
-          val itemLayout = inflater.inflate(R.layout.list_patientsummary_prescription_drug, parent, false)
-          return DrugViewHolder(itemLayout)
-        }
-      }
-    }
   }
 }
